Clarify key autorepeat behaviour
There exists client that send repeated "down" events on automatic key
repeats, and those that send pairs of "up" and "down". Usually this
behaviour depends on the platform on which the client runs.

Since only the repeated "down" case allows the server to tell the
difference between automatic and manual key repeats, clarify that the
"down" repeat method is the preferred.

Also clarify that the client should handle auto repeat because of
latency problems.

Signed-off-by: Pierre Ossman <ossman@cendio.se>
Signed-off-by: Peter Rosin <peda@lysator.liu.se>