Clarify extended desktop size & vmware display mode interaction
The VMware display mode change psuedo encoding is intended to allow the
server to notify the client of a change in both desktop dimensions and
pixel format at the same time, such that the client has a serialization
point at which to safely adapt to changes in remote desktop native
pixel format and dimensions.

A client that is interested in multi-head layout information, however,
cannot use the new dimensions provided with the vmware display mode
change rectangle.

To enable the clients to retain ability to atomically adapt to changes
in pixel format and desktop layout, the server needs to be written
such that it sends a extended desktop size rectangle immediately
following any VMware display mode change rectangle.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>