Clarify Open H.264 handling of oversized frames
This clarifies how clients should treat frames that are either larger or
smaller than the rectangle to which they belong.

Hardware encoders are often limited in that frame sizes need to be
multiples of certain numbers such as 16 or 32. It is therefor
advantageous to allow servers to submit oversized frames, expecting the
client to crop them.

Signed-off-by: Andri Yngvason <andri@yngvason.is>