Clarify colour map handling with VMWare Display Mode Change
The SetPixelFormat message allows for a client initiated pixel format
change. It defines that the color map is empty immediately after the
message is received. The server is expected to send a SetColorMapEntries
message before any further non-psuedo encodings, if true-color-flag was
false in the new pixel format.

The VMware Display Mode Change psuedo-encoding allows for a server
initiated pixel format change, concurrently with a framebuffer resize.
This change suggests that the colour map should similarly be defined
as empty after this point, with the server expected to send a
SetColorMapEntries message before any further non-psuedo encodings,
if true-color-flag was false.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>